# sih-backend/contracts/openapi.yaml
openapi: "3.0.3"
info:
  title: Unigraph — Real-Time Crypto Fraud Attribution API
  description: |
    Backend API for Unigraph (SIH26183 — MHA/I4C).

    **Auth schemes:**
    - `bearerAuth` (JWT): all `/api/v1/...` routes except `/api/v1/auth/login` and `/api/v1/auth/refresh`
    - `apiKeyAuth` (X-API-Key header): `/check-wallet` only — used by external VASPs, never by the frontend JWT client

    **Error envelope** (all failures, every endpoint):
    ```json
    { "error": { "code": "string", "message": "string", "details": {} } }
    ```

    **Enums** are defined in `contracts/entities.md` and must never be extended inline.
  version: "0.1.0"
  contact:
    name: Backend Owner (Rehan)
servers:
  - url: http://localhost:8000
    description: Local development

# ── Security schemes ─────────────────────────────────────────────────────────
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    apiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key

  # ── Shared schemas ──────────────────────────────────────────────────────────
  schemas:
...
    CaseStatus:
      type: string
      enum: [new, investigating, escalated_to_vasp, frozen, closed]
      description: "Closed enum — matches Kanban columns exactly. Do not extend."
    AlertAction:
      type: string
      enum: [allow, hold, block]
      description: "Closed enum — returned by /check-wallet and stored on Alert records."
...
    TriggeredBy:
      type: string
      enum: [check_wallet_hook, registry_refresh, manual]
...
    ErrorEnvelope:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
              example: WALLET_NOT_FOUND
            message:
              type: string
              example: "No wallet found for that address/chain pair."
            details:
              type: object
              additionalProperties: true
              example: { "address": "1A1zP1...", "chain": "BTC" }

    # ── Pagination wrapper ────────────────────────────────────────────────────
    PaginationMeta:
      type: object
      required: [total, page, page_size]
      properties:
        total:
          type: integer
          example: 142
        page:
          type: integer
          example: 1
        page_size:
          type: integer
          example: 25

    # ── Core entities ─────────────────────────────────────────────────────────
...
    Case:
      type: object
      required: [id, status, opened_at]
      properties:
        id:
          type: string
          format: uuid
          example: "a1b2c3d4-0000-4000-8000-000000000001"
        status:
          $ref: "#/components/schemas/CaseStatus"
        assigned_investigator:
          type: string
          nullable: true
          example: "inspector_sharma"
        opened_at:
          type: string
          format: date-time
          example: "2026-08-26T08:00:00Z"
        closed_at:
          type: string
          format: date-time
          nullable: true
          example: null
    Alert:
      type: object
      required: [id, wallet_id, triggered_by, action, created_at]
      properties:
        id:
          type: string
          format: uuid
          example: "d4e5f6a7-1111-4111-b111-111111111111"
        wallet_id:
          type: string
          format: uuid
          example: "550e8400-e29b-41d4-a716-446655440000"
        case_id:
          type: string
          format: uuid
          nullable: true
          example: "a1b2c3d4-0000-4000-8000-000000000001"
        triggered_by:
          $ref: "#/components/schemas/TriggeredBy"
        action:
          $ref: "#/components/schemas/AlertAction"
        created_at:
          type: string
          format: date-time
          example: "2026-08-28T07:30:00Z"
        resolved_at:
          type: string
          format: date-time
          nullable: true
          example: null

    # ── Auth schemas ──────────────────────────────────────────────────────────
    LoginRequest:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
          example: "investigator@i4c.gov.in"
        password:
          type: string
          format: password
          example: "s3cr3t"
    LoginResponse:
      type: object
      required: [access_token, refresh_token, role]
      properties:
        access_token:
          type: string
          example: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
        refresh_token:
          type: string
          example: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
        role:
          type: string
          enum: [admin, investigator, compliance_viewer]
          example: "investigator"
    RefreshRequest:
      type: object
      required: [refresh_token]
      properties:
        refresh_token:
          type: string
          example: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
    RefreshResponse:
      type: object
      required: [access_token]
      properties:
        access_token:
          type: string
          example: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."

    # ── Request bodies ────────────────────────────────────────────────────────
...
    CaseCreate:
      type: object
      properties:
        assigned_investigator:
          type: string
          nullable: true
          example: "inspector_sharma"
        wallet_ids:
          type: array
          items:
            type: string
            format: uuid
        initial_status:
          $ref: "#/components/schemas/CaseStatus"
          default: "new"
    CasePatch:
      type: object
      description: "At least one field must be provided."
      properties:
        status:
          $ref: "#/components/schemas/CaseStatus"
        assigned_investigator:
          type: string
          nullable: true
          example: "inspector_sharma"
...
  responses:
    Unauthorized:
      description: Missing or invalid JWT token
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
          example:
            error:
              code: UNAUTHORIZED
              message: "Missing or invalid authentication token."
              details: {}
    Forbidden:
      description: Authenticated but insufficient role
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
          example:
            error:
              code: FORBIDDEN
              message: "Your role does not permit this action."
              details: {}
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
    UnprocessableEntity:
      description: Validation error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"

# ── Global default auth (JWT bearer) ─────────────────────────────────────────
security:
  - bearerAuth: []

# ── Paths ─────────────────────────────────────────────────────────────────────
paths:
...
  /api/v1/auth/login:
    post:
      summary: Obtain JWT access + refresh tokens
      operationId: auth_login
      security: [] # no auth — this IS the login
      tags: [auth]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: Tokens issued
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LoginResponse"
        "401":
          description: Bad credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
              example:
                error:
                  code: INVALID_CREDENTIALS
                  message: "Email or password is incorrect."
                  details: {}
        "422":
          $ref: "#/components/responses/UnprocessableEntity"
  /api/v1/auth/refresh:
    post:
      summary: Exchange a refresh token for a new access token
      operationId: auth_refresh
      security:
        - bearerAuth: [] # refresh token sent as Bearer
      tags: [auth]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RefreshRequest"
      responses:
        "200":
          description: New access token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RefreshResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "422":
          $ref: "#/components/responses/UnprocessableEntity"

  # ── Complaints ─────────────────────────────────────────────────────────────
...
  /api/v1/cases:
    post:
      summary: Create a new investigation case
      operationId: create_case
      tags: [cases]
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CaseCreate"
      responses:
        "201":
          description: Created case
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Case"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "422":
          $ref: "#/components/responses/UnprocessableEntity"

    get:
      summary: List cases with optional status filter
      operationId: list_cases
      tags: [cases]
      security:
        - bearerAuth: []
      parameters:
        - name: status
          in: query
          schema:
            $ref: "#/components/schemas/CaseStatus"
        - name: page
          in: query
          schema:
            type: integer
            default: 1
            minimum: 1
        - name: page_size
          in: query
          schema:
            type: integer
            default: 25
            minimum: 1
            maximum: 100
      responses:
        "200":
          description: Paginated case list
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/PaginationMeta"
                  - type: object
                    required: [items]
                    properties:
                      items:
                        type: array
                        items:
                          $ref: "#/components/schemas/Case"
              example:
                total: 17
                page: 1
                page_size: 25
                items:
                  - id: "a1b2c3d4-0000-4000-8000-000000000001"
                    status: "investigating"
                    assigned_investigator: "inspector_sharma"
                    opened_at: "2026-08-26T08:00:00Z"
                    closed_at: null
        "401":
          $ref: "#/components/responses/Unauthorized"
  /api/v1/cases/{id}:
    get:
      summary: Get case details by ID
      operationId: get_case
      tags: [cases]
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Case details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Case"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"

    patch:
      summary: Update case status or assigned investigator
      operationId: update_case
      tags: [cases]
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CasePatch"
      responses:
        "200":
          description: Updated case
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Case"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "422":
          $ref: "#/components/responses/UnprocessableEntity"
  /api/v1/cases/{id}/report:
    get:
      summary: Download PDF case report
      operationId: get_case_report
      tags: [cases]
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: PDF report binary
          content:
            application/pdf:
              schema:
                type: string
                format: binary
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"

  # ── Alerts ────────────────────────────────────────────────────────────────
  /api/v1/alerts:
    get:
      summary: List alerts (polled every 5–10s by the Alerts screen)
      operationId: list_alerts
      tags: [alerts]
      security:
        - bearerAuth: []
      parameters:
        - name: resolved
          in: query
          schema:
            type: boolean
          description: "Filter: true = resolved only, false = open only, omit = all"
        - name: page
          in: query
          schema:
            type: integer
            default: 1
            minimum: 1
        - name: page_size
          in: query
          schema:
            type: integer
            default: 25
            minimum: 1
            maximum: 100
      responses:
        "200":
          description: Paginated alert list
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/PaginationMeta"
                  - type: object
                    required: [items]
                    properties:
                      items:
                        type: array
                        items:
                          $ref: "#/components/schemas/Alert"
              example:
                total: 8
                page: 1
                page_size: 25
                items:
                  - id: "d4e5f6a7-1111-4111-b111-111111111111"
                    wallet_id: "550e8400-e29b-41d4-a716-446655440000"
                    case_id: "a1b2c3d4-0000-4000-8000-000000000001"
                    triggered_by: "check_wallet_hook"
                    action: "hold"
                    created_at: "2026-08-28T07:30:00Z"
                    resolved_at: null
        "401":
          $ref: "#/components/responses/Unauthorized"
...